+++ test/multi-select/multi-select.component.spec.js
@@ -1,0 +1,35 @@
+/* eslint-disable prefer-arrow-callback */
+/* eslint-disable react/jsx-filename-extension */
+
+import React from 'react';
+import { expect } from 'chai';
+import { mount } from 'enzyme';
+import { Map } from 'immutable';
+
+import { MultiSelect } from '../../src/multi-select/index';
+
+
+describe('MultiSelect component', function describe() {
+  it('should render correctly', function it() {
+    const props = {
+      checkedItems: Map(),
+      items: [
+        {
+          id: 1,
+          text: 'Item 1',
+        },
+        {
+          id: 2,
+          text: 'Item 2',
+        },
+      ],
+    };
+
+    const wrapper = mount(
+      <MultiSelect {...props} />,
+    );
+
+    expect(wrapper.find('.oc-multi-select-item-checkbox').length).to.eql(2);
+    expect(wrapper.find('.oc-multi-select-item-text').at(0).text()).to.eql('Item 1');
+  });
+});